requirement.txt
langchain==0.2.1
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.22.0
faiss-cpu==1.7.4
streamlit-extras
 